requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
webdriver-manager>=4.0.0
lxml>=4.9.0
colorama>=0.4.6
tabulate>=0.9.0
streamlit>=1.28.0
plotly>=5.17.0
python-docx>=1.1.0
# LLM dependencies (optionnelles - installez seulement celles dont vous avez besoin)
# openai>=1.0.0  # Pour OpenAI API
# mistralai>=1.0.0  # Pour Mistral AI
# anthropic>=0.18.0  # Pour Anthropic Claude
# requests est déjà installé pour Ollama
# Extraction PDF pour le CV
PyPDF2>=3.0.0
# Airflow pour l'automatisation (optionnel - installez seulement si vous utilisez Airflow)
# apache-airflow>=2.7.0  # Pour l'automatisation quotidienne

